feat: migrated   4 sections from the  components user my penas join penas and the game list
